strain map updates
